requirements.txt
pytest>=6.0.0
pytest-json-report>=1.5.0
gitpython>=3.1.18
pydantic>=1.9.0
colorama>=0.4.4
